pyproject.toml
[tool.poetry]
name = "rsrg"
version = "0.1.0"
description = "FDK scrawler"
authors = ["Erich Raschle <erich.raschle@rsrg.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Scrapy = "^2.7.0"
ipython = "^8.5.0"
scrapy-splash = "^0.8.0"
Pillow = "^9.2.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"
mypy = "^0.982"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [project]
# name = "my_package"
# description = "My package description"
# readme = "README.rst"
# requires-python = ">=3.7"
# keywords = ["one", "two"]
# license = { text = "BSD 3-Clause License" }
# classifiers = ["Framework :: Django", "Programming Language :: Python :: 3"]
# dependencies = ["requests", 'importlib-metadata; python_version<"3.8"']
# dynamic = ["version"]

# [project.optional-dependencies]
# pdf = ["ReportLab>=1.2", "RXP"]
# rest = ["docutils>=0.3", "pack ==1.1, ==1.3"]

# [project.scripts]
# my-script = "my_package.module:function"

# ... other project metadata fields as specified in:
#     https://packaging.python.org/en/latest/specifications/declaring-project-metadata/

[tool.isort]
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
